requirements.txt
numpy
streamlit
opencv-python
scikit-learn
lightgbm
catboost
joblib
matplotlib
seaborn
pillow
